[bootstrap] Commission Delegated Regulation (EU) 2017/2055 of 23 June 2017 supplementing Directive (EU) 2015/2366 of the European Parliament and of the Council with regard to regulatory technical standards for the cooperation and exchange of information between competent authorities relating to the exercise of the right of establishment and the freedom to provide services of payment institutions (Text with EEA relevance. ) — versión original 2025
Publicación original de Commission Delegated Regulation (EU) 2017/2055 of 23 June 2017 supplementing Directive (EU) 2015/2366 of the European Parliament and of the Council with regard to regulatory technical standards for the cooperation and exchange of information between competent authorities relating to the exercise of the right of establishment and the freedom to provide services of payment institutions (Text with EEA relevance. ).

Norma: 32017R2055
Fecha: 2025-02-20
Fuente: http://data.europa.eu/eli/reg_del/2017/2055/oj

Source-Id: 32017R2055
Source-Date: 2025-02-20
Norm-Id: 32017R2055